Restructure messaging module
