--- tg_bot/services/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ("send_message",)
-
-from .send_message import send_message
