+++ tg_bot/utils/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ("create_media_group",)
+
+from .media import create_media_group
